oid-yolo/requirements.txt
boto3~=1.19.12
botocore~=1.22.12
matplotlib~=3.4.3
opencv-python~=4.5.4.58
pandas~=1.3.4
PyYAML~=6.0
tqdm~=4.62.3
